+++ src/gitlab_migrator/__init__.py
@@ -1,3 +1,3 @@
 """GitLab Group Migrator."""
 
-__version__ = "0.7.0"
+__version__ = "0.7.1"